move barrier insertion pass to passes.mapping
