+++ qiskit/transpiler/passes/__init__.py
@@ -11,12 +11,12 @@ from .cx_cancellation import CXCancellation
 from .fixed_point import FixedPoint
 from .optimize_1q_gates import Optimize1qGates
 from .decompose import Decompose
-from .barrier_before_final_measurements import BarrierBeforeFinalMeasurements
+from .commutation_analysis import CommutationAnalysis
+from .commutation_transformation import CommutationTransformation
+from .mapping.barrier_before_final_measurements import BarrierBeforeFinalMeasurements
 from .mapping.check_map import CheckMap
-from .mapping.stochastic_swap import StochasticSwap
 from .mapping.cx_direction import CXDirection
 from .mapping.unroller import Unroller
 from .mapping.basic_swap import BasicSwap
 from .mapping.lookahead_swap import LookaheadSwap
-from .commutation_analysis import CommutationAnalysis
-from .commutation_transformation import CommutationTransformation
+from .mapping.stochastic_swap import StochasticSwap
